refactor: 레거시 클래스를 삭제하라
레거시 클래스를 삭제하고 의존성을 수정한다.

BREAKING CHANGE:
	변경 사항:
	- lotto.Lotto을 삭제한다.
	- lotto.LottoMachine을 삭제한다.
	- 삭제한 클래스와의 의존성을 처리한다.

+++ src/main/java/lotto/domain/LottoPrize.java
@@ -10,7 +10,7 @@ public enum LottoPrize {
 
     public final long value;
 
-    LottoPrize(int value) {
+    LottoPrize(long value) {
         this.value = value;
     }
 }
